Fix : 카테고리 색 선택 뷰 파일로 분리
