+++ Cookiee/Cookiee/Presentation/MyPage/Category/View/CategoryAddView.swift
@@ -136,30 +136,3 @@ struct CategoryAddView: View {
         }
     }
 }
-
-struct ColorPickerBottomSheetView: View {
-    @Binding var isPresented: Bool
-    @Binding var selectedColor: String
-    
-    var body: some View {
-        VStack {
-            Color.clear
-                .contentShape(Rectangle())
-                .onTapGesture {
-                    withAnimation {
-                        isPresented = false
-                    }
-                }
-            ZStack {
-                UIColorPickerViewControllerWrapper(selectedColor: $selectedColor) {
-                    withAnimation {
-                        isPresented = false
-                    }
-                }
-                .frame(height: 520)
-            }
-            .background(Color.white)
-            .zIndex(1)
-        }
-    }
-}
